Separate out command station related code to it's own project
