+++ src/TauStellwerk.CommandStations/ECoS/ECoSMessageDecoder.cs
