+++ src/TauStellwerk.Server/Services/EngineControlService/ActiveEngine.cs
@@ -4,6 +4,7 @@
 // </copyright>
 
 using TauStellwerk.Base;
+using TauStellwerk.CommandStations;
 using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Server.Services;
