+++ src/TauStellwerk.Server/Services/EngineControlService/EngineStateManager.cs
@@ -6,6 +6,7 @@
 using System.Collections.Concurrent;
 using FluentResults;
 using TauStellwerk.Base;
+using TauStellwerk.CommandStations;
 using TauStellwerk.Data.Model;
 
 namespace TauStellwerk.Server.Services;
