+++ src/TauStellwerk.Server/Startup.cs
@@ -9,10 +9,10 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Options;
 using TauStellwerk.Base;
+using TauStellwerk.CommandStations;
 using TauStellwerk.Data;
 using TauStellwerk.Data.Dao;
 using TauStellwerk.Data.ImageService;
-using TauStellwerk.Server.CommandStations;
 using TauStellwerk.Server.Hub;
 using TauStellwerk.Server.Services;
